requirements.txt
pygame==2.1.2
python-dotenv==1.0.0
